refactor: moved index to misc
